+++ api/models/Batch.go
@@ -9,7 +9,6 @@ type Batch struct {
 	Dispatched  bool   `gorm:"default:false"`
 	Summary     int    `gorm:"not null" json:"summary"`
 	BatchUserID uint32 `sql:"type:int REFERENCES users(id)" json:"batch_user_id"`
-	BatchUser   User   `json:"batch_user"`
 }
 type Filter struct {
 	UserID int64 `schema:"userid"`
